Refactor: Modify findPosts, findPost query
