+++ routes/index.js
@@ -1,6 +1,6 @@
 const { Router } = require("express");
 const router = Router();
-const notice = require("./notice");
+const notice = require("./board");
 
 router.use("/notice", notice);
